Prepared backend for register and login page (unit tests passed), database created locally (postgresql on docker). Tested by unit tests, postmand and dbeaver.

+++ backend/backend/src/main/java/org/example/backend/repository/UserRepository.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    // Możesz dodać dodatkowe metody, np. wyszukiwanie po nazwie użytkownika lub emailu
-    Optional<User> findByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email); // Znajdź użytkownika po emailu (np. do logowania)
 }
